Moved AliasTab components to its own folder
