+++ lib/screens/alias_tab/alias_tab.dart
@@ -1,10 +1,10 @@
 import 'package:anonaddy/global_providers.dart';
 import 'package:anonaddy/models/alias/alias_model.dart';
+import 'package:anonaddy/screens/alias_tab/components/alias_shimmer_loading.dart';
+import 'package:anonaddy/screens/alias_tab/components/alias_tab_pie_chart.dart';
 import 'package:anonaddy/shared_components/constants/material_constants.dart';
 import 'package:anonaddy/shared_components/list_tiles/alias_list_tile.dart';
 import 'package:anonaddy/shared_components/lottie_widget.dart';
-import 'package:anonaddy/shared_components/pie_chart/alias_tab_pie_chart.dart';
-import 'package:anonaddy/shared_components/shimmer_effects/alias_shimmer_loading.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
